requirements.txt
Django==5.1
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
flasgger==0.9.7.1
Flask==3.0.3
Flask-Caching==2.3.0
mysqlclient==2.2.4
redis==5.0.8
setuptools==73.0.0
